requirements.txt
nose==1.3.0
nose_xunitmp==0.2
pep8==1.4.6
requests==2.4.2
retrying==1.3.1
websocket-client==0.40.0
